requirements.txt
PyMySQL
Appium-Python-Client
selenium
PyYAML
uiautomator2
tqdm
loguru